修正了BUG, 添加了新功能
